requirements.txt
argostranslate==1.9.6
colorama==0.4.6
dill==0.3.9
moviepy==2.1.2
numpy==2.2.6
ollama==0.5.1
opencv_python==4.11.0.86
opencv_python_headless==4.11.0.86
Pillow==11.2.1
pyttsx3==2.98
tqdm==4.67.1
translate==3.6.1
